CodingAssistant/pyproject.toml
[project]
name = "codingassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "docx>=0.2.4",
    "fastapi>=0.110.0",
    "openai>=1.82.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
codingassistant = "codingassistant:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
